+++ messages.hpp
@@ -33,8 +33,8 @@ namespace messages
    void encode_set_tlc_ch(uint8_t* p, uint8_t  ch, uint16_t  value);
    void decode_set_tlc_ch(uint8_t* p, uint8_t &ch, uint16_t &value);
 
-   void encode_start_effect(uint8_t* p, uint8_t  effect_id, uint32_t  start_time, uint16_t  duration);
-   void decode_start_effect(uint8_t* p, uint8_t &effect_id, uint32_t &start_time, uint16_t &duration);
+   void encode_start_effect(uint8_t* p, uint8_t  effect_id, uint32_t  start_time, uint32_t  duration);
+   void decode_start_effect(uint8_t* p, uint8_t &effect_id, uint32_t &start_time, uint32_t &duration);
 
    void decode_status(uint8_t* p, uint16_t &slave_id, uint32_t &t_rx, int8_t &major_version, int8_t &minor_version,
                       uint16_t &vcell, uint16_t &soc, uint16_t &missed_message_count, uint8_t &freshness_count);
